fix: Refactor DailyViewAdapter to SlideViewAdapter
