+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/WriteViewFragment.java
@@ -19,7 +19,6 @@ import java.time.temporal.ChronoUnit;
 import snu.swpp.moment.LoginRegisterActivity;
 import snu.swpp.moment.MainActivity;
 import snu.swpp.moment.data.repository.AuthenticationRepository;
-import snu.swpp.moment.ui.main_writeview.DaySlide.DailyViewAdapter;
 import snu.swpp.moment.databinding.FragmentWriteviewBinding;
 import snu.swpp.moment.utils.TimeConverter;
 
@@ -58,7 +57,7 @@ public class WriteViewFragment extends Fragment {
         // 여기부터 slide가 가능해짐
         //ViewPager2
         mPager = binding.viewpager;
-        pagerAdapter = new DailyViewAdapter(WriteViewFragment.this, num_page);
+        pagerAdapter = new SlideViewAdapter(WriteViewFragment.this, num_page);
         mPager.setAdapter(pagerAdapter);
         mPager.setOrientation(ViewPager2.ORIENTATION_HORIZONTAL);
         // 최초의 페이지 설정은 이거로함 (numpage 보다 크면 마지막페이지가 세팅되는듯 - 아래 onPageChangeCallBack)
